Added Player Leave Message
When a player leaves, a message will now pop up, stating the name of the player that left.
Found a difficult bug; when joining a game via IP, and leaving, it would softlock the client.
This ment that the client couldn't create a game, find games, they could join via ip, but not log out anymore.
This has been fixed.