Progress making Turn Based
Currently making it Turn Based, this means a lot of talking between player and server, this isn't as easy as I thought.
Currently the player can move, go to the next player, let them move but it breaks for some reason, so it goes to the next player again.
Next push should hopefully be that everyone can move.