TradeTile Bases - Broken
Added the Base for the TradeTile.
This works as long as a player doesn't have to do anything.
I tried to give the player an option, as placeholder just to roll, and when that has ended, they would buy the tile.
This doesn't work yet.

After some brainstorming, might be best to forward all tile logic to the PlayerController, and handle it like that.