Working prototype Tile Server interaction
Finally got a Server Interaction Prototype working.

When a player lands on a TradeTile (just a normal building), a Case will be made.
This will be send to an event on the Local PlayerController.
Based on this, the Player will make local stuff (UI) and is able to manipulate it.

When this is done, the player will call another local event, which will then forward everything to a Server call with that TradeTile.

Again with the Case, the TradeTile will handle it accordingly.
Once done, the game resumes play.