Added way to pay
When landing on a tile owned by someone else, you will see the info of the tile.
There will be text stating which row is active. (how many tiles of that type the other player has)
You can then roll to pay a certain amount.
Currently nothing will happen when you roll a 6, game will softlock if you can't pay up.

Noticed a big bug, when setting a tile actor to hidden in game, it will break the game in some way.
To hide the tile, we need to hide the exact cube in the actor, then it works fine.