Added way to pay and steal
Added a pay the other players or steal their trade tile.
All logic works, but there is 1 more visual bug in the PayScreen.
For some reason if the server player buys the tradetile, the correct text to pay will not show on a remote client.
This push is for safety, will now try to fix that.