Modified everything + Added PlayerStates
I added PlayerStates to keep track of all data that should be known by everyone.
This is something I should've done from the start, because now I had to modify ever BP to accept PlayerStates instead of PlayerControllers.

This allows clients to see the data of other clients.
I came across the problem when I was trying to tell the player who had to pay which row was active. This is based of how many properties that owner has that are that type.
Because this was stored in the PlayerController, you couldn't see that.
I noticed that if you couldn't see that, there might be a lot more problems later on, with diploma's etc.

TLDR; PlayerStates good, confused me at first so took a while to get it working.