Added Duel Template + Implemented Dalton City
Added a template that will be used for Dueling, this is currently just an event in the Tile that takes 2 players states.
Once activated correct things should be handled, but this is for later.

Implemented Dalton City.
This is a special one, the street exists of 2 lanes, yet if you land on a space opposite of a building, you still have the chance to buy that building. This also means that dueling should also happen when you land next to another player.

This is done by adding Overlapping Tiles.
These first check if another player is standing on the other tile, if there is, duel.
If there isn't, we execute the tiles, depending on which one has a valid component.