Modified the Payment Handling
Modified the Payment Handling of BP_TradeTile.
Now a function get's called to avoid duplicate code.
This makes it also a lot easier to keep track of everything.
The function just needs the player that needs their money removed, the player that should get paid, and the amount.

If the player cannot pay the amount, bankruptcy can get called. (still needs to be implemented)