Added Bankruptcy
Whenever a player doesn't have enough money, they will go bankrupt.
All leftover money and trade tiles will go to the player that needed to receive money.
Disclaimer; this is not how it is in the actual game.
In the actual game, you have an option to sell cards for 50% of their purchase value.
This might get added later on, but since the focus for now is just Trade Tiles, this works fine.