Added Game Loop
Whenever a player bankrupts, the Game State gets notified and keeps track of this.
When all but 1 player is bankrupt, the game state will switch Match State to EndGame.
Then it will get the name of the last surviving player, and display a new EndScreen widget.
The host has the option to go back to the lobby here. (would be nice to add loading screen for this too)

- Bug found, dueling in Dalton City is currently broken, need to figure more out about this.